docs: no-progress only works in non-interactive mode
It confused me that no-progress didn't work in the interactive mode.

After reading the source code, I realized that it only works in
non-interactive mode.

This makes sense because it seems that no-progress is useful only when
its output is used in a script.

So I believe this is not a bug but the help text needs to mention this.