Make the bar graph meaningful in large directories
The current bar implementation shows an empty graph in a directory with
approximately 20 files (not so uncommon) of similar size - each file
size is 5% of the total, which rounds to 0 for the bar of 10 ticks.
With directories with more items, the usefulness of the bar degrades
even faster even for objects of different sizes.

With this change, all sizes are shown relative to the largest
item which gives a better idea of how big the items are.